Completed all business job-management related items
