+++ bin/templates/ajax_job_detail_table.html
@@ -1,0 +1,7 @@
+<ul class="collection with-header">
+    <!--<li class="collection-header">Title:<b> </b></li>-->
+    {% for key, value in single_job_detail.items() %}
+      <li class="collection-item">{{key}}: <b>{{value}}</b></li>
+    {% endfor %}
+</ul>
+
