+++ bin/views_ajax.py
@@ -5,7 +5,7 @@ from bin import app, mysql, do_sms
 @app.route("/ajax/ajax_business_get_jobs", methods=['POST'])
 def ajax_business_get_jobs():
     bus_id = request.get_json().get('bus_id','')
-    types = request.get_json().get('types','')
+    types = request.get_json().get('types','').split(",")
     
     jobs = []
     
@@ -18,5 +18,26 @@ def ajax_business_get_jobs():
         if len(data) is not 0:
             jobs.extend(data)
     
-    return jsonify(render_template('ajax/job_row.html',
+    if len(jobs) is 0:
+        return jsonify("<tr><td>No Closed Jobs</td></tr>")
+    
+    return jsonify(render_template('ajax_business_job_rows.html',
                            jobs=jobs))
+
+@app.route("/ajax/ajax_job_detail_table", methods=['POST'])
+def ajax_job_detail_table():
+    id_job = request.get_json().get('id_job','')
+
+    cursor = mysql.connection.cursor()
+    cursor.callproc('get_job',[id_job])
+    data = cursor.fetchall()
+    cursor.close()
+        
+    if len(data) is not 0:
+        job = data[0]
+    
+    return jsonify({'id_job':id_job,
+                    'table_html':
+                    render_template('ajax_job_detail_table.html',
+                           single_job_detail=job)
+                    })
