+++ web/css/dispatcher.css
@@ -14,6 +14,6 @@
 .promo i{margin:40px 0;color:#004d40;font-size:7rem;display:block}
 .promo-caption{font-size:1.7rem;font-weight:500;margin-top:5px;margin-bottom:0}
 
-table.highlight-yellow tbody tr:hover {
+table.highlight-tr tbody tr:hover {
     background-color: #b2dfdb;
 }
